+++ slick_reporting/__init__.py
@@ -1,5 +1,5 @@
 default_app_config = "slick_reporting.apps.ReportAppConfig"
 
-VERSION = (1, 1, 1)
+VERSION = (1, 2, 0)
 
-__version__ = "1.1.1"
+__version__ = "1.2.0"